chore: update status pattern (#4110)
* chore: update status pattern

* chore: feedback from pr

* chore: more feedback

* chore: more feedback again

* chore: another round of feedback and name change
